Consolidate XHR UIExplorer examples
