+++ Examples/UIExplorer/js/UIExplorerList.ios.js
@@ -354,7 +354,7 @@ const APIExamples: Array<UIExplorerExample> = [
   },
   {
     key: 'XHRExample',
-    module: require('./XHRExample.ios'),
+    module: require('./XHRExample'),
     supportsTVOS: true,
   },
 ];
